Refactored "Anagrafica and "RAppresentanteFiscaleCessionario" to have common things in abstract class "AbstractNaming"
